+++ PoGoEmulator/PoGoEmulator/PoGoEmulator.csproj
@@ -202,6 +202,7 @@
       <HintPath>$(SolutionDir)\packages\System.Text.Encodings.Web.4.3.1\lib\netstandard1.0\System.Text.Encodings.Web.dll</HintPath>
     </Reference>
     <Reference Include="System.Transactions" />
+    <Reference Include="System.Web" />
     <Reference Include="System.Windows" />
     <Reference Include="System.Xml.Linq" />
     <Reference Include="System.Data.DataSetExtensions" />
@@ -242,6 +243,7 @@
     <Compile Include="Requests\RequestHandler.cs" />
     <Compile Include="Models\TcpServer.cs" />
     <Compile Include="Requests\RpcRequestParser.cs" />
+    <Compile Include="Responses\Response.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
